autoppia_iwa copy/requirements.txt
beautifulsoup4==4.12.3
dataclasses_json==0.6.7
datasets==2.20.0
markdownify==0.12.1
matplotlib==3.9.3
networkx==3.3
numpy
orjson==3.10.12
playwright==1.49.0
pydantic==2.10.4
pymongo==4.10.1
python-dotenv==1.0.1
Requests==2.32.3
screeninfo==0.8.1
setuptools
spacy==3.7.5
tiktoken==0.7.0
webdriver_manager==4.0.1
openai==1.58.1
selenium==4.27.1
pytest==8.3.4
dependency-injector==4.41.0
pymongo==4.10.1
flask==3.1.0
lxml==5.3.1
pre-commit==4.1.0
pillow==11.1.0
